+++ src/pixelator/graph/utils.py
@@ -148,6 +148,7 @@ def create_node_markers_counts(
     graph: Graph,
     k: int = 0,
     normalization: Optional[Literal["mean"]] = None,
+    name_as_index: bool = True,
 ) -> pd.DataFrame:
     """Create a matrix of marker counts for each in the graph.
 
@@ -180,16 +181,17 @@ def create_node_markers_counts(
     node_marker_counts = pd.DataFrame.from_records(
         list(graph.vs.get_attribute("markers")),
         columns=markers,
-        index=list(graph.vs.get_attribute("name")),
+        index=list(graph.raw.nodes),
     )
     node_marker_counts = node_marker_counts.reindex(
         sorted(node_marker_counts.columns), axis=1
     )
     node_marker_counts.columns.name = "markers"
     node_marker_counts.columns = node_marker_counts.columns.astype("string[pyarrow]")
-    node_marker_counts.index = pd.Index(
-        list(graph.vs.get_attribute("name")), dtype="string[pyarrow]", name="node"
-    )
+    if name_as_index:
+        node_marker_counts.index = pd.Index(
+            list(graph.vs.get_attribute("name")), dtype="string[pyarrow]", name="node"
+        )
     if k == 0:
         return node_marker_counts